minimega: spring cleaning in winter
Removed dead code and variables. Removed vyatta API.
